+++ lib/runners/changes.rb
@@ -32,10 +32,11 @@ module Runners
       gdp = patches # NOTE: This assignment is required for typecheck by Steep
       if gdp
         location = issue.location
-        return true unless location
         patch = gdp.find_patch_by_file(issue.path.to_s)
         if patch && location
           patch.changed_lines.one? { |line| location.start_line == line.number }
+        elsif patch
+          true
         else
           false
         end