+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 [Full diff](https://github.com/sider/runners/compare/0.22.3...HEAD)
 
 - Fix #patches and use ... for git-diff(1) [#955](https://github.com/sider/runners/pull/955)
+- Fix Changes#include? to return false for unchanged files [#954](https://github.com/sider/runners/pull/954)
 
 ## 0.22.3